Added GlobalCounter POC and cleaned up the issues that exposed

+++ src/Portal/UI/ViewModel/IViewModel.cs
@@ -6,6 +6,6 @@ namespace Eventually.Portal.UI.ViewModel
     {
         Guid Id { get; set; }
 
-        int Version { get; set; }
+        long Version { get; set; }
     }
 }
